backend/requirements.txt
flask
python-dotenv
google-generativeai
flask-cors
pypdf
langchain
chromadb
tiktoken
langchain-community
langchain-google-genai

